chore: move cache dict implementation to shared dir
